Split and comment  database+account management
